Moved component from index.js to a separate file
